+++ R/round.r
@@ -47,21 +47,31 @@
 #'
 #' @rdname round_date
 #' @param x a vector of date-time objects
-#' @param unit a character string specifying a time unit or a multiple of a unit
-#'   to be rounded to. Valid base units are `second`, `minute`, `hour`, `day`,
-#'   `week`, `month`, `bimonth`, `quarter`, `season`, `halfyear` and
-#'   `year`. Arbitrary unique English abbreviations as in the [period()]
-#'   constructor are allowed. Rounding to multiples of units (except weeks) is
-#'   supported.
-#' @param change_on_boundary if this is `NULL` (the default), instants on the boundary
-#'   remain unchanged, but `Date` objects are rounded up to the next boundary.
-#'   If this is `TRUE`, instants on the boundary are rounded up to the next boundary.
-#'   If this is `FALSE`, nothing on the boundary is rounded up at all. This was the
-#'   default for \pkg{lubridate} prior to `v1.6.0`.
-#'   See section `Rounding Up Date Objects` below for more details.
+#' @param unit a string or a date-time object. When a singleton string, it
+#'   specifies a time unit or a multiple of a unit to be rounded to. Valid base
+#'   units are `second`, `minute`, `hour`, `day`, `week`, `month`, `bimonth`,
+#'   `quarter`, `season`, `halfyear` and `year`. Arbitrary unique English
+#'   abbreviations as in the [period()] constructor are allowed. Rounding to
+#'   multiples of units (except weeks) is supported.
+#'
+#'   When `unit` is a date-time object rounding is done to the nearest of the
+#'   elements in `unit`. If range of `unit` vector does not cover the range of
+#'   `x` `ceiling_date()` and `floor_date()` round to the `max(x)` and `min(x)`
+#'   for elements that fall outside of `range(unit)`.
+#' @param change_on_boundary if this is `NULL` (the default), instants on the
+#'   boundary remain unchanged, but `Date` objects are rounded up to the next
+#'   boundary. If this is `TRUE`, instants on the boundary are rounded up to the
+#'   next boundary. If this is `FALSE`, nothing on the boundary is rounded up at
+#'   all. This was the default for \pkg{lubridate} prior to `v1.6.0`. See
+#'   section `Rounding Up Date Objects` below for more details.
 #'
 #' @param week_start when unit is `week`, specify the reference day.
 #'   7 represents Sunday and 1 represents Monday.
+#'
+#' @return When `unit` is a string, return a Date object if `x` is a Date and
+#'   `unit` is larger or equal than "day", otherwise a POSIXct object. When
+#'   `unit` is a date-time object, return a date-time object of the same class
+#'   and same time zone as `unit`.
 #' @keywords manip chron
 #' @seealso [base::round()]
 #' @examples
@@ -132,32 +142,50 @@
 #' ceiling_date(ymd_hms("2000-01-01 00:00:00"), "month", change_on_boundary = FALSE)
 #' ceiling_date(ymd("2000-01-01"), "month")
 #'
-
 #' @export
 round_date <- function(x, unit = "second", week_start = getOption("lubridate.week.start", 7)) {
 
-  if (!length(x)) return(x)
+  if (length(x) == 0)
+    return(x)
 
-  parsed_unit <- parse_period_unit(unit)
-  n <- parsed_unit$n
-  basic_unit <- standardise_period_names(parsed_unit$unit)
+  xx <- as_datetime(x, tz = tz(x))
 
-  new <-
-    if (n == 1 && basic_unit %in% c("second", "minute", "hour", "day")) {
-      ## special case for fast rounding
-      round.POSIXt(as_datetime(x, tz = tz(x)), units = lub2base_units[[basic_unit]])
-    } else {
-      above <- unclass(as.POSIXct(ceiling_date(x, unit = unit, week_start = week_start)))
-      mid <- unclass(as_datetime(x, tz = tz(x)))
-      below <- unclass(as.POSIXct(floor_date(x, unit = unit, week_start = week_start)))
-      wabove <- (above - mid) <= (mid - below)
-      wabove <- !is.na(wabove) & wabove
-      new <- below
-      new[wabove] <- above[wabove]
-      .POSIXct(new, tz = tz(x))
-    }
+  .round <- function(below, above) {
+    mid <- unclass(xx)
+    below <- unclass(below)
+    above <- unclass(above)
+    wabove <- (above - mid) <= (mid - below)
+    wabove <- !is.na(wabove) & wabove
+    new <- below
+    new[wabove] <- above[wabove]
+    .POSIXct(new, tz = tz(x))
+  }
 
-  reclass_date_maybe(new, x, unit)
+  if (is.POSIXct(unit) || is.Date(unit)) {
+
+    u <- as_datetime(unit, tz = tz(unit))
+    below <- floor_to_posix(xx, u, TRUE)
+    above <- ceil_to_posix(xx, u, TRUE)
+    reclass_date(.round(below, above), unit)
+
+  } else {
+
+    parsed_unit <- parse_period_unit(unit)
+    n <- parsed_unit$n
+    basic_unit <- standardise_period_names(parsed_unit$unit)
+
+    new <-
+      if (n == 1 && basic_unit %in% c("second", "minute", "hour", "day")) {
+        ## special case for fast rounding
+        round.POSIXt(xx, units = lub2base_units[[basic_unit]])
+      } else {
+        below <- as.POSIXct(floor_date(xx, unit = unit, week_start = week_start))
+        above <- as.POSIXct(ceiling_date(xx, unit = unit, week_start = week_start))
+        .round(below, above)
+      }
+    reclass_date_maybe(new, x, unit)
+
+  }
 }
 
 reclass_date_maybe <- function(new, orig, unit) {
@@ -171,7 +199,15 @@ reclass_date_maybe <- function(new, orig, unit) {
 #' @rdname round_date
 #' @export
 floor_date <- function(x, unit = "seconds", week_start = getOption("lubridate.week.start", 7)) {
-  if (!length(x)) return(x)
+  if (length(x) == 0)
+    return(x)
+
+  xx <- as_datetime(x, tz = tz(x))
+
+  if (is.POSIXct(unit) || is.Date(unit)) {
+    new <- floor_to_posix(xx, as_datetime(unit, tz = tz(unit)))
+    return(reclass_date(new, unit))
+  }
 
   parsed_unit <- parse_period_unit(unit)
   n <- parsed_unit$n
@@ -180,7 +216,7 @@ floor_date <- function(x, unit = "seconds", week_start = getOption("lubridate.we
   if (unit %in% c("second", "minute", "hour", "day")) {
 
     # as_datetime is necesary for correct tz = UTC when x is Date
-    out <- trunc_multi_unit(as_datetime(x, tz = tz(x)), unit, n)
+    out <- trunc_multi_unit(xx, unit, n)
     reclass_date_maybe(out, x, unit)
 
   } else {
@@ -194,31 +230,34 @@ floor_date <- function(x, unit = "seconds", week_start = getOption("lubridate.we
        (n > 1 && unit == "month")) {
       new_months <-
         switch(unit,
-               month    = floor_multi_unit1(month(x), n),
-               bimonth  = floor_multi_unit1(month(x), 2 * n),
-               quarter  = floor_multi_unit1(month(x), 3 * n),
-               halfyear = floor_multi_unit1(month(x), 6 * n),
-               season   = floor_multi_unit(month(x), 3 * n))
+               month    = floor_multi_unit1(month(xx), n),
+               bimonth  = floor_multi_unit1(month(xx), 2 * n),
+               quarter  = floor_multi_unit1(month(xx), 3 * n),
+               halfyear = floor_multi_unit1(month(xx), 6 * n),
+               season   = floor_multi_unit(month(xx), 3 * n))
       n <- Inf
       unit <- "month"
     }
 
-    switch(unit,
-           week     = update(x, wdays = 1, hours = 0, minutes = 0, seconds = 0, week_start = week_start),
-           month    = {
-             if (n > 1) update(x, months = new_months, mdays = 1, hours = 0, minutes = 0, seconds = 0)
-             else      update(x, mdays = 1, hours = 0, minutes = 0, seconds = 0)
-           },
-           year     = {
-             ## due to bug https://github.com/tidyverse/lubridate/issues/319 we
-             ## need to do it in two steps
-             if (n > 1) {
-               y <- update(x, ydays = 1, hours = 0, minutes = 0, seconds = 0)
-               update(y, years = floor_multi_unit(year(y), n))
-             } else {
-               update(x, ydays = 1, hours = 0, minutes = 0, seconds = 0)
-             }
-           })
+    new <-
+      switch(unit,
+        week     = update(xx, wdays = 1, hours = 0, minutes = 0, seconds = 0, week_start = week_start),
+        month    = {
+          if (n > 1) update(xx, months = new_months, mdays = 1, hours = 0, minutes = 0, seconds = 0)
+          else      update(xx, mdays = 1, hours = 0, minutes = 0, seconds = 0)
+        },
+        year     = {
+          ## due to bug https://github.com/tidyverse/lubridate/issues/319 we
+          ## need to do it in two steps
+          if (n > 1) {
+            y <- update(xx, ydays = 1, hours = 0, minutes = 0, seconds = 0)
+            update(y, years = floor_multi_unit(year(y), n))
+          } else {
+            update(xx, ydays = 1, hours = 0, minutes = 0, seconds = 0)
+          }
+        })
+
+    reclass_date_maybe(new, x, unit)
   }
 }
 
@@ -240,8 +279,15 @@ floor_date <- function(x, unit = "seconds", week_start = getOption("lubridate.we
 #'  ceiling_date(x, "month", change_on_boundary = TRUE)
 ceiling_date <- function(x, unit = "seconds", change_on_boundary = NULL, week_start = getOption("lubridate.week.start", 7)) {
 
-  if (!length(x))
+  if (length(x) == 0)
     return(x)
+
+  xx <- as_datetime(x, tz = tz(x))
+
+  if (is.POSIXct(unit) || is.Date(unit)) {
+    new <- ceil_to_posix(xx, as_datetime(unit, tz = tz(unit)))
+    return(reclass_date(new, unit))
+  }
 
   parsed_unit <- parse_period_unit(unit)
   n <- parsed_unit$n
@@ -266,14 +312,13 @@ ceiling_date <- function(x, unit = "seconds", change_on_boundary = NULL, week_st
   } else if (unit %in% c("minute", "hour")) {
 
     ## as_datetime converts Date to POSIXct with tz=UTC
-    new <- as_datetime(x, tz = tz(x))
     delta <- switch(unit, minute = 60, hour = 3600, day = 86400) * n
     new <-
       if (change_on_boundary) {
-        trunc_multi_unit(new, unit, n) + delta
+        trunc_multi_unit(xx, unit, n) + delta
       } else {
-        new1 <- trunc_multi_unit(new, unit, n)
-        not_same <- which(new1 != new)
+        new1 <- trunc_multi_unit(xx, unit, n)
+        not_same <- which(new1 != xx)
         new1[not_same] <- new1[not_same] + delta
         new1
       }
@@ -287,8 +332,8 @@ ceiling_date <- function(x, unit = "seconds", change_on_boundary = NULL, week_st
 
     ## need this to accomodate the case when date is on a boundary
     new <-
-      if (change_on_boundary) x
-      else update(x, seconds = second(x) - 0.00001, simple = T)
+      if (change_on_boundary) xx
+      else update(xx, seconds = second(xx) - 0.00001, simple = T)
 
     if (unit %in% c("month", "bimonth", "quarter", "halfyear", "season")) {
       new_month <-
@@ -358,4 +403,47 @@ ceil_multi_unit <- function(x, n) {
 
 ceil_multi_unit1 <- function(x, n) {
   (((x - 1) %/% n) *  n) + n + 1L
+}
+
+check_round_endpoints <- function(unit) {
+  if (length(unit) == 0)
+    stop("Zero length `unit` argument", call. = FALSE)
+  if (any(is.na(unit)))
+    stop("NAs in `unit` argument`", call. = FALSE)
+}
+
+floor_to_posix <- function(x, endpoints, tight = FALSE) {
+  check_round_endpoints(endpoints)
+  if (all(is.na(x))) return(x)
+  endpoints <- sort(endpoints)
+  ux <- unclass(x)
+  min <- min(ux, na.rm = TRUE)
+  breaks <- unclass(endpoints)
+  emin <- breaks[[1]]
+  if (min < emin) {
+    beg <- .POSIXct(if (tight) emin else min, tz = tz(endpoints))
+    endpoints <- c(beg, endpoints)
+    breaks <- c(min, breaks)
+  }
+  breaks <- c(breaks, Inf)
+  bins <- .bincode(ux, breaks, right = FALSE, include.lowest = TRUE)
+  endpoints[bins]
+}
+
+ceil_to_posix <- function(x, endpoints, tight = FALSE) {
+  check_round_endpoints(endpoints)
+  if (all(is.na(x))) return(x)
+  endpoints <- sort(endpoints)
+  ux <- unclass(x)
+  max <- max(ux, na.rm = TRUE)
+  breaks <- unclass(endpoints)
+  emax <- breaks[[length(breaks)]]
+  if (emax < max) {
+    end <- .POSIXct(if (tight) emax else max, tz = tz(endpoints))
+    endpoints <- c(endpoints, end)
+    breaks <- c(breaks, max)
+  }
+  breaks <- c(-Inf, breaks)
+  bins <- .bincode(ux, breaks, right = TRUE, include.lowest = TRUE)
+  out <- endpoints[bins]
 }